requirements.txt
Flask
gunicorn
langchain
langchain-community
langchain-google-genai
langchain-chroma
python-dotenv
rapidocr-onnxruntime
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime